refactor pagination

+++ apis/books/backend/flaskr/__init__.py
@@ -1,7 +1,7 @@
 import os
 import sys
 from flask import Flask, request, abort, jsonify
-from flask_sqlalchemy import SQLAlchemy  # , or_
+from flask_sqlalchemy import SQLAlchemy
 from flask_cors import CORS
 import random
 
@@ -15,6 +15,15 @@ BOOKS_PER_SHELF = 8
 #   - Make sure for each route that you're thinking through when to abort and with which kind of error
 #   - If you change any of the response body keys, make sure you update the frontend to correspond.
 
+def paginate_books (request, books):
+    page = request.args.get('page', 1, type=int)
+    start = (page -1) * BOOKS_PER_SHELF
+    end = start + BOOKS_PER_SHELF
+
+    books = [book.format() for book in books]
+    books_on_page = books[start:end]
+
+    return books_on_page
 
 def create_app(test_config=None):
     # create and configure the app
@@ -25,12 +34,8 @@ def create_app(test_config=None):
     # CORS Headers
     @app.after_request
     def after_request(response):
-        response.headers.add(
-            "Access-Control-Allow-Headers", "Content-Type,Authorization,true"
-        )
-        response.headers.add(
-            "Access-Control-Allow-Methods", "GET,PATCH,POST,DELETE,OPTIONS"
-        )
+        response.headers["Access-Control-Allow-Headers"] = "Content-Type,Authorization,true"
+        response.headers["Access-Control-Allow-Methods"] = "GET,PATCH,POST,DELETE,OPTIONS"
         return response
 
     # @TODO: Write a route that retrivies all books, paginated.
@@ -41,19 +46,29 @@ def create_app(test_config=None):
     # TEST: When completed, the webpage will display books including title, author, and rating shown as stars
     @app.route('/books', methods=['GET'])
     def get_books():
-        page = request.args.get('page', 1, type=int)
-        start = (page - 1) * BOOKS_PER_SHELF
-        end = start + BOOKS_PER_SHELF
-        try:
-            books = Book.query.all()
-            formatted_books = [book.format() for book in books]
-        except:
-            print(sys.exc_info())
+        books = Book.query.order_by(Book.id).all()
+        books_on_page = paginate_books(request, books)
+        if len(books_on_page) == 0:
+            abort(404)
         return jsonify({
             'success': True,
-            'books': formatted_books[start:end],
-            'total_books': len(formatted_books)
+            'books': books_on_page,
+            'total_books': len(books_on_page)
         })
+        # page = request.args.get('page', 1, type=int)
+        # start = (page - 1) * BOOKS_PER_SHELF
+        # end = start + BOOKS_PER_SHELF
+        # try:
+        #     books = Book.query.all()
+        #     formatted_books = [book.format() for book in books]
+        # except:
+        #     print(sys.exc_info())
+        #     abort(404)
+        # return jsonify({
+        #     'success': True,
+        #     'books': formatted_books[start:end],
+        #     'total_books': len(formatted_books)
+        # })
     # @TODO: Write a route that will update a single book's rating.
     #         It should only be able to update the rating, not the entire representation
     #         and should follow API design principles regarding method and route.
@@ -88,21 +103,36 @@ def create_app(test_config=None):
     #        Response body keys: 'success', 'books' and 'total_books'
     @app.route('/books/<int:book_id>', methods=['DELETE'])
     def delete_book(book_id):
-        book = Book.query.filter(book_id == Book.id).one_or_none()
-        book.delete()
-        page = request.args.get('page', 1, type=int)
-        start = (page - 1) * BOOKS_PER_SHELF
-        end = start + BOOKS_PER_SHELF
         try:
-            books = Book.query.all()
-            formatted_books = [book.format() for book in books]
+            book = Book.query.filter(Book.id == book_id).one_or_none()
+            if book is None:
+                abort(404)
+            book.delete()
+            books = Book.query.order_by(Book.id).all()
+            books_on_page = paginate_books(request, books)
+            return jsonify({
+                'success': True,
+                'deleted': book_id,
+                'books': books_on_page,
+                'total_books': len(books_on_page)
+            })
         except:
-            print(sys.exc_info())
-        return jsonify({
-            'success': True,
-            'books': formatted_books[start:end],
-            'total_books': len(formatted_books)
-        })
+            abort(422)
+        # book = Book.query.filter(book_id == Book.id).one_or_none()
+        # book.delete()
+        # page = request.args.get('page', 1, type=int)
+        # start = (page - 1) * BOOKS_PER_SHELF
+        # end = start + BOOKS_PER_SHELF
+        # try:
+        #     books = Book.query.all()
+        #     formatted_books = [book.format() for book in books]
+        # except:
+        #     print(sys.exc_info())
+        # return jsonify({
+        #     'success': True,
+        #     'books': formatted_books[start:end],
+        #     'total_books': len(formatted_books)
+        # })
     # TEST: When completed, you will be able to delete a single book by clicking on the trashcan.
 
     # @TODO: Write a route that create a new book.
@@ -118,31 +148,23 @@ def create_app(test_config=None):
             rating = body['rating']
         )
         try:
-            print(book)
             book.insert()
-            page = request.args.get('page', 1, type=int)
-            start = (page - 1) * BOOKS_PER_SHELF
-            end = start + BOOKS_PER_SHELF
-            try:
-                books = Book.query.all()
-                formatted_books = [book.format() for book in books]
-            except:
-                print(sys.exc_info())
-            return jsonify({
+            books = Book.query.order_by(Book.id).all()
+            books_on_page = paginate_books(request, books)
+            return ({
                 'success': True,
                 'created': book.id,
-                'books': formatted_books[start:end],
-                'total_books': len(formatted_books)
+                'books': books_on_page,
+                'total_books': len(books_on_page)
             })
         except:
-            print(sys.exc_info())
-            abort(422)
+            abort(422)            
 
     @app.errorhandler(400)
     def bad_request(error):
         return jsonify({
             'success': False,
-            'error': error,
+            'error': '400',
             'message': 'could not process bad request'
         }), 400
     
@@ -150,15 +172,23 @@ def create_app(test_config=None):
     def not_found(error):
         return jsonify({
             'success': False,
-            'error': error,
+            'error': '404',
             'message': 'resource not found'
         }), 404
+
+    @app.errorhandler(405)
+    def not_allowed(error):
+        return jsonify({
+            'success': False,
+            'error': '405',
+            'message': 'method not allowed'
+        })
 
     @app.errorhandler(422)
     def unprocessable(error):
         return jsonify({
             'success': False,
-            'error': error,
+            'error': '422',
             'message': 'resource unprocessable'
         }), 422
